refactor(language & message): move language and message stuff around
