+++ src/lib/core/index.ts
@@ -1,6 +1,7 @@
 export * from './module';
 
 export * from './activity';
+export * from './language';
 export * from './media';
 export * from './message';
 export * from './request';
