+++ src/lib/core/language/shared/index.ts
@@ -1,4 +1,3 @@
+export * from './language.loader';
 export * from './language.service';
-export * from './language.service.provider';
-export * from './language-loader';
 export * from './missing-translation.guard';
