+++ src/lib/core/language/shared/language.service.ts
@@ -1,5 +1,23 @@
+import { Http } from '@angular/http';
 import { Injectable } from '@angular/core';
-import { TranslateService } from '@ngx-translate/core';
+import { TranslateService, TranslateLoader } from '@ngx-translate/core';
+
+import { LanguageLoader } from './language.loader';
+
+
+export function defaultLanguageLoader(http: Http) {
+  return new LanguageLoader();
+}
+
+export function provideLanguageLoader(loader?) {
+
+  return {
+    provide: TranslateLoader,
+    useFactory: (loader) || (defaultLanguageLoader),
+    deps: [Http]
+  };
+}
+
 
 @Injectable()
 export class LanguageService {
