+++ src/lib/core/message/shared/index.ts
@@ -1,0 +1,3 @@
+export * from './message.enum';
+export * from './message.interface';
+export * from './message.service';
