+++ src/lib/core/message/shared/message.enum.ts
