+++ src/lib/core/message/shared/message.interface.ts
