--- src/lib/language/index.ts
@@ -1,3 +1,0 @@
-export * from './module';
-
-export * from './shared';
